+++ src/Resources/contao/languages/de/tl_mitgliederverwaltung_applications.php
@@ -3,10 +3,13 @@
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['application_legend'] = 'Name';
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['application_legend'] = 'Bewerbung';
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['tournament'] = array('Turnier', 'Wählen Sie das Turnier aus, für das sich der Spieler beworben hat.');
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['applicationDate'] = array('Bewerbungsdatum', 'Bewerbungsdatum im Format TT.MM.JJJJ');
+
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['promise_legend'] = 'Zusage';
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['promiseDate'] = array('Zusagedatum', 'Zusagedatum im Format TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['comment'] = array('Kommentar', 'Interner Kommentar');
 
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['publish_legend'] = 'Aktivieren';
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_applications']['published'] = array('Aktiv', 'Aktivieren oder deaktivieren Sie hier das Turnier');